+++ CHANGELOG.md
@@ -34,6 +34,7 @@ developers to maintain and readjust their custom modifications on the main proje
 - Make the phone number field of backend users optional (#1709)
 - Change the main CRUD layout and navigation structure (list first)
 - Replace the availabilities type with the new slot interval field
+- Keep the provider option selected in the booking page if possible (#1672)
 
 ### Fixed